ssh-keyscan returns extra line
`ssh-keyscan` has started returning an extra line (a comment) in the
following pattern:

```
«host-or-IP» ssh-ed25519 «public key data»
```

This interferes with the operation of the `cut` that is needed to trim
the `«host-or-IP»` from the second line which, in turn, is necessary
for the public key to be treated as valid input during certificate
generation.

The fix is to include the `-q` option to silence output of the comment.

Signed-off-by: Phill <34226495+Paraphraser@users.noreply.github.com>